[feat] Dynamic Vault Auth Token Injection
This commit adds dockerfile to dockerise the plugin, adds pod.yaml to
deploy the plugin as a pod in kube-system namespace, adds required files
for vault kubernetes authentication and vault agent which will be
responsible for authenticating with vault and writing the token into the
pod dynamically before the vault kms plugin pod boots up. Also updates
the readme to include necessary steps with the latest changes made to
code and deployments.